Ensure code base is at v3
A repo is allowed to host more than one module version, the recommended
way to do so is to have versioned subdirs. But that is not the only way,
tagged versions and branches is fine too. In either case the import
paths need to match to version specified in go.mod.

Since this repo is obviously not doing the versioned sub-dir route the
import paths of all top-level files/first-level directories/packages
need to be updated to specify they are for v3. Otherwise it seems like
the repo is in a weird mixed version mode.